# docs/swagger.yaml
basePath: ./
definitions:
  dto.Brand:
    properties:
      id:
        type: integer
      name:
        type: string
      status_id:
        type: integer
    type: object
  dto.BrandUpdate:
    properties:
      id:
        type: integer
      name:
        type: string
      status_id:
        type: integer
    type: object
  dto.CategoriesTree:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CategoryTree'
        type: array
    type: object
  dto.Category:
    properties:
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      sequence:
        type: integer
      status_id:
        type: integer
    type: object
  dto.CategoryTree:
    properties:
      category_name:
        type: string
      children:
        items:
          $ref: '#/definitions/dto.CategoryTree'
        type: array
      id:
        type: integer
    type: object
  dto.CategoryUpdate:
    properties:
      id:
        type: integer
      name:
        type: string
      sequence:
        type: integer
      status_id:
        type: integer
    type: object
  dto.Error:
    properties:
      message:
        type: string
    type: object
  dto.IDWrapper:
    properties:
      id:
        minimum: 1
        type: integer
    required:
    - id
    type: object
  dto.PaginatedBrandCollection:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.Brand'
        type: array
      total:
        description: This will always return the total of all records
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Describes technoStore REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: technoStore API
  version: "1.0"
paths:
  /v1/brand:
    post:
      consumes:
      - application/json
      description: Create a new Brand in the system
      parameters:
      - description: Brand params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Brand'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.IDWrapper'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Add a new Brand
      tags:
      - Brand
  /v1/brand/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Brand by id
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Brand delete processed
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Brand not found
          schema:
            type: string
        "500":
          description: Error
          schema:
            type: string
      summary: Delete a Brand by id
      tags:
      - Brand
    get:
      consumes:
      - application/json
      description: Get a Brand by id
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Brand'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Get a Brand by id
      tags:
      - Brand
    patch:
      consumes:
      - application/json
      description: Update a Brand by id
      parameters:
      - description: Brand params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BrandUpdate'
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: BrandDto updated
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Update a Brand by id
      tags:
      - Brand
  /v1/brands:
    get:
      consumes:
      - application/json
      description: Get Brands
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedBrandCollection'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Error
          schema:
            type: string
      summary: Get Brands
      tags:
      - Brand
  /v1/categories:
    get:
      consumes:
      - application/json
      description: Get categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoriesTree'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Error
          schema:
            type: string
      summary: Get categories
      tags:
      - Category
  /v1/category:
    post:
      consumes:
      - application/json
      description: Create a new Category in the system
      parameters:
      - description: Category params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.IDWrapper'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Add a new Category
      tags:
      - Category
  /v1/category/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Category delete processed
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Category not found
          schema:
            type: string
        "500":
          description: Error
          schema:
            type: string
      summary: Delete a Category by id
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Get a Category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Category'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Get a Category by id
      tags:
      - Category
    patch:
      consumes:
      - application/json
      description: Update a Category by id
      parameters:
      - description: Category params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryUpdate'
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: CategoryDto updated
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Update a Category by id
      tags:
      - Category
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"